fix(news): aplicar restrição de nulo e ajustar tipo numérico nas entidades
- Adiciona a anotação @Column(nullable = false) ao campo `authorId`
  na entidade News para garantir consistência no banco de dados.
- Altera a tipagem do atributo `order` em ContentBlock de `int` para
  `Long` para evitar problemas de limite numérico ou padronizar integrações.

+++ src/main/java/com/ccomp/br/domain/news/persistence/ContentBlock.java
@@ -1,7 +1,7 @@
 package com.ccomp.br.domain.news.persistence;
 
 public record ContentBlock(
-        int order,
+        Long order,
         String type,
         String content,
         String url,
